punctuator/requirements.txt
torch
numpy
transformers
attrs
